various cleanup
