+++ nomadic/core/__init__.py
@@ -2,29 +2,18 @@ import html
 from nomadic.core.models import Note, Notebook
 from nomadic.core.search import search, search_pdf
 
-def _process(text, escape=False):
-    text = text.decode('utf-8')
-    if escape:
-        return html.escape(text)
-    return text
 
 class Nomadic():
     def __init__(self, notes_path):
         self.notes_path = notes_path
         self.rootbook = Notebook(notes_path)
 
-    def search(self, query, delimiters=('<b>','</b>'), window=150, include_pdf=False):
+    def search(self, query, delimiters=('<b>','</b>'), window=150, include_pdf=False, html_out=False):
         """search across txt/md and pdf files
         window -> num characters to show before/after match
         delimiters -> what to surround matches with
+        html_out -> whether or not output will be to html
         """
-
-        # check if the delimiters look like html delimiters
-        html_delim = True
-        for d in delimiters:
-            if not(d.startswith('<') and d.endswith('>')):
-                html_delim = False
-
         results = []
         for note_path, matches in search(query).items():
             note = Note(note_path)
@@ -34,11 +23,11 @@ class Nomadic():
                     frm = max(0, start - window)
                     to = min(len(text), start + end + window)
                     snippet = \
-                        _process(text[frm:start], escape=html_delim) + \
+                        self._process(text[frm:start], escape=html_out) + \
                         delimiters[0] + \
-                        _process(text[start:start+end], escape=html_delim) + \
+                        self._process(text[start:start+end], escape=html_out) + \
                         delimiters[1] + \
-                        _process(text[start+end:to], escape=html_delim)
+                        self._process(text[start+end:to], escape=html_out)
 
                     if frm > 0:
                         snippet = '...{}'.format(snippet)
@@ -53,3 +42,7 @@ class Nomadic():
                 note = Note(note_path)
                 results.append((note, matches))
         return results
+
+    def _process(text, escape=False):
+        text = text.decode('utf-8')
+        return html.escape(text) if escape else text
