+++ nomadic/server/assets/templates/export/default.html
@@ -1,7 +1,6 @@
 {% extends 'layout.html' %}
 
 {% block content %}
-
     <style>
         html, body {
             padding: 0 1em;
@@ -68,5 +67,4 @@
         });
         MathJax.Hub.Startup.onload();
     </script>
-
 {% endblock %}
