+++ nomadic/server/assets/templates/export/presentation.html
@@ -1,7 +1,6 @@
 {% extends 'layout.html' %}
 
 {% block content %}
-
     <style>
         html, body {
             padding: 0 1em;
@@ -150,5 +149,4 @@
         });
         MathJax.Hub.Startup.onload();
     </script>
-
 {% endblock %}
