+++ nomadic/server/assets/templates/notebook.html
@@ -33,6 +33,3 @@
     <div class="notes--list">{{ render_notebook(notebook) }}</div>
   </div>
 {% endblock %}
-
-{% block scripts %}
-{% endblock %}
